src/data_ingest_requirements.txt
google-cloud-pubsub==2.18.4
pydantic==2.3.0
python-dotenv==1.0.0
tinydb==4.8.0
